code_capnhattrangthaithanhtoan

+++ database/migrations/2021_11_11_144500_add_column_status_to_table_orders.php
@@ -4,29 +4,23 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-class AddColumnStatusToTableOrders extends Migration
-{
-    /**
-     * Run the migrations.
-     *
-     * @return void
-     */
+return new class extends Migration {
     public function up()
     {
         Schema::table('orders', function (Blueprint $table) {
-            $table->integer('status')->default(0);
+            if (!Schema::hasColumn('orders', 'status')) { // Kiểm tra trước khi thêm
+                $table->integer('status')->default(0);
+            }
         });
     }
 
-    /**
-     * Reverse the migrations.
-     *
-     * @return void
-     */
     public function down()
     {
         Schema::table('orders', function (Blueprint $table) {
-            $table->dropColumn('status');
+            if (Schema::hasColumn('orders', 'status')) {
+                $table->dropColumn('status');
+            }
         });
     }
-}
+};
+
